requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
resend>=2.0.0
twilio>=9.0.0
